fix (styles): make styles work in build
