+++ .parcelrc
@@ -1,11 +1,11 @@
 {
   "extends": "@parcel/config-webextension",
   "transformers": {
-    "*.svg": [
-      "@parcel/transformer-svg-react"
-    ],
     "*.css.ts": [
       "@vanilla-extract/parcel-transformer"
+    ],
+    "*.svg": [
+      "@parcel/transformer-svg-react"
     ],
     "*.{js,mjs,jsx,cjs,ts,tsx}": [
       "@parcel/transformer-js",
