+++ src/popup/Popup.tsx
@@ -3,8 +3,7 @@ import { DataProvider } from './providers/DataProvider';
 import { RouterProvider } from './providers/RouterProvider';
 import { Router } from './screens/Router';
 
-// TODO: is it neccessary to import it here?
-// import '../../Popup.css';
+import './Popup.css';
 
 export default function Popup() {
   return (
