+++ src/popup/popup.common.html
@@ -1,12 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <!-- build:style -->
-    <title>Clean the mess! addon</title>
-  </head>
-  <body>
-    <main id="main-container"></main>
-    <script type="module" src="./Popup.tsx"></script>
-  </body>
+
+<head>
+  <meta charset="utf-8">
+  <!-- build:style -->
+  <title>Clean the mess! addon</title>
+</head>
+
+<body>
+  <main id="main-container"></main>
+  <script type="module" src="./Popup.tsx"></script>
+</body>
+
 </html>
