+++ src/popup/popup.firefox.html
@@ -1,3 +1,3 @@
 <!-- build:style -->
-<link rel="stylesheet" href="../styles/firefox/style.scss">
+<!-- <link rel="stylesheet" href="../styles/firefox/style.scss"> -->
 <!-- /build:style -->
