+++ src/popup/screens/Search/SearchScreen.tsx
@@ -1,7 +1,7 @@
 import { FC, useMemo } from 'react';
 import { DetailsItem } from '../../components/DetailItem';
 import { useFavicons, useData, useSearch } from '../../hooks';
-import { screenBodyContainer } from '../../Popup.css';
+import { screenBodyContainer } from '../Router.css';
 import { screenList } from '../Overview/OverviewScreen.css';
 import { SearchError } from './SearchError';
 import { SearchHeader } from './SearchHeader';
